refactor: User와 DeliveryDriver를 각각 독립적인 애그리거트로 분리
Resolved #118
