--- user/src/main/java/com/business/user/deliverydriver/infrastructure/client/AuthClient.java
@@ -1,4 +1,0 @@
-package com.business.user.infrastructure.client;
-
-public interface AuthClient {
-}
