+++ user/src/main/java/com/business/user/user/infrastructure/client/AuthClient.java
@@ -1,0 +1,4 @@
+package com.business.user.user.infrastructure.client;
+
+public interface AuthClient {
+}
